Fixed not loading bug
